Small Changes to update functions
Pair Swap, WorldRowManager to fixed update to reduce amount of times its called, as player can only move so fast we don't have to check their location that often.
SingleRowObstSpanwer similar thing, we are running a for loop inside the update function for as many rows are on the screen. So I changed it to fixed update, if anything we can also change it to  even slower?